+++ Runtime/API/V2/Models/ApiChatHistory.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using Neocortex.Data;
+
+namespace Neocortex.API
+{
+    [Serializable]
+    public class ApiChatHistory
+    {
+        public string sessionId;
+        public Message[] messages;
+    }
+}