+++ Runtime/Data/Enums/ApiResponseDataType.cs
@@ -1,9 +1,0 @@
-namespace Neocortex.Data
-{
-    public enum ApiResponseDataType
-    {
-        Text,
-        Audio,
-        Texture
-    }
-}